core-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
cryptography==41.0.7
python-multipart==0.0.6
pyyaml==6.0.1
bcrypt>=4.0.1,<4.1.0
python-dotenv==1.2.1
sqlalchemy==2.0.23
passlib==1.7.4
alembic==1.12.1
psycopg2-binary==2.9.9
pytest==7.4.0
pytest-asyncio==0.22.0
httpx==0.24.1
aiosqlite==0.19.0
# Optional: Redis for production (verification codes, webhook idempotency, caching)
# redis[asyncio]==5.0.0  # Uncomment if using Redis in production


